+++ platforms/esp32/CMakeLists.txt
@@ -1,11 +1,8 @@
 cmake_minimum_required(VERSION 3.16)
 
-enable_language(C)
-enable_language(CXX)
-
 set(CMAKE_SYSTEM_PROCESSOR "${TARGET}")
 #set(IDF_TOOLCHAIN clang)
 
-project(husarnet_esp32_core)
+include($ENV{IDF_PATH}/tools/cmake/project.cmake)
 
-include(${CMAKE_SOURCE_DIR}/../../core/husarnet.cmake)
+project(husarnet_esp32_core)
