+++ platforms/esp32/components/husarnet_core/CMakeLists.txt
@@ -1,0 +1,4 @@
+# System name does not propagate from toolchain file in first script mode pass
+set(CMAKE_SYSTEM_NAME Generic)
+
+include(${CMAKE_SOURCE_DIR}/../../core/husarnet.cmake)
