+++ platforms/esp32/components/husarnet_core/sdkconfig
@@ -1,0 +1,1 @@
+CONFIG_NEWLIB_NANO_FORMAT=y
